+++ IsekaiMod/Content/Features/IsekaiProtagonist/Archetypes/EdgeLord/EdgeLordProficiencies.cs
@@ -13,8 +13,9 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.EdgeLord {
         public static void Add() {
             var EdgeLordProficiencies = Helpers.CreateBlueprint<BlueprintFeature>(IsekaiContext, "EdgeLordProficiencies", bp => {
                 bp.SetName(IsekaiContext, "Edge Lord Proficiencies");
-                bp.SetDescription(IsekaiContext, "The Edge Lord is proficient with all simple and martial weapons as well as light and medium armor. They can cast {g|Encyclopedia:Spell}spells{/g} from "
-                    + "this class while wearing armor without incurring the normal {g|Encyclopedia:Spell_Fail_Chance}arcane spell failure chance{/g}, but they incur the normal arcane spell "
+                bp.SetDescription(IsekaiContext, "The Edge Lord is proficient with all simple, martial, and exotic weapons as well as light and medium armor. "
+                    + "They can cast {g|Encyclopedia:Spell}spells{/g} from this class while wearing armor without incurring the normal "
+                    + "{g|Encyclopedia:Spell_Fail_Chance}arcane spell failure chance{/g}, but they incur the normal arcane spell "
                     + "failure chance for arcane spells received from other classes.");
                 bp.AddComponent<AddFacts>(c => {
                     c.m_Facts = new BlueprintUnitFactReference[] {
@@ -22,6 +23,7 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.Archetypes.EdgeLord {
                         StaticReferences.Proficiencies.MediumArmorProficiency.ToReference<BlueprintUnitFactReference>(),
                         StaticReferences.Proficiencies.SimpleWeaponProficiency.ToReference<BlueprintUnitFactReference>(),
                         StaticReferences.Proficiencies.MartialWeaponProficiency.ToReference<BlueprintUnitFactReference>(),
+                        StaticReferences.Proficiencies.ExoticWeaponProficiency.ToReference<BlueprintUnitFactReference>(),
                     };
                 });
                 bp.AddComponent<ArcaneArmorProficiency>(c => {